Move SonarLintConfiguration to the correct namespace
